Renames login to singin since login is used to represent the voter login
