+++ pages/signin/index.tsx
@@ -6,7 +6,7 @@ import i18n from '../../i18n'
 
 import { Column, Grid } from '../../components/grid'
 import { PageCard } from '../../components/cards'
-import { LogInForm, LogInImport } from '../../components/login'
+import { SignInForm, SignInImport } from '../../components/signin'
 import { Account } from '../../lib/types'
 import { useDbAccounts } from '../../hooks/use-db-accounts'
 import { useWallet } from '../../hooks/use-wallet'
@@ -14,7 +14,7 @@ import { useMessageAlert } from '../../hooks/message-alert'
 import { useResponsive } from '../../hooks/use-window-size'
 import { DASHBOARD_PATH } from '../../const/routes'
 
-const LogInPage = () => {
+const SignInPage = () => {
   const { dbAccounts } = useDbAccounts()
   const { laptop } = useResponsive()
   const { restoreEncryptedWallet } = useWallet()
@@ -31,7 +31,7 @@ const LogInPage = () => {
         account.hdPath,
         passphrase
       )
-      
+
       router.push(DASHBOARD_PATH)
     } catch (error) {
       setAlertMessage(i18n.t('sign_in.invalid_passphrase'))
@@ -43,7 +43,7 @@ const LogInPage = () => {
       <Grid>
         {hasAccounts ? (
           <Column lg={colSmSize}>
-            <LogInForm
+            <SignInForm
               accounts={dbAccounts}
               onSubmit={handlerSubmit}
             />
@@ -53,7 +53,7 @@ const LogInPage = () => {
         ) : null}
 
         <Column lg={colSmSize}>
-          <LogInImport hasAccount={hasAccounts} />
+          <SignInImport hasAccount={hasAccounts} />
         </Column>
       </Grid>
     </PageCard>
@@ -66,4 +66,4 @@ const LoginDivider = styled.div`
   border-bottom: solid 2px ${({ theme }) => theme.lightBorder };
 `
 
-export default LogInPage
+export default SignInPage
